requirements.txt
mcp>=1.0.0
spotipy>=2.23.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
